% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cm-fit.R
\name{cm_assoc_inference}
\alias{cm_assoc_inference}
\title{Format conditional-model association inference}
\usage{
cm_assoc_inference(par, Vyth, assoc_idx)
}
\arguments{
\item{par}{Optimized conditional-model parameter vector.}

\item{Vyth}{Variance-covariance matrix.}

\item{assoc_idx}{Indexes of the two association parameters.}
}
\value{
Numeric vector with association estimates, standard errors, and log
p-values.
}
\description{
Format conditional-model association inference
}
\keyword{internal}
