% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cm-fit.R
\name{cM_devfun_xifix_datl}
\alias{cM_devfun_xifix_datl}
\title{Build fixed-xi deviance functions from prepared data}
\usage{
cM_devfun_xifix_datl(lxi0 = 0, datl, Sfun, lc = lc_fun())
}
\arguments{
\item{lxi0}{Fixed log xi value.}

\item{datl}{Prepared conditional-model data list.}

\item{Sfun}{List of Weibull distribution helper functions.}

\item{lc}{Censored-subject integration function.}
}
\value{
A list of functions used by \code{\link[=fixxi_fitcm0]{fixxi_fitcm0()}} and
\code{\link[=fixxi_fitcm_noB]{fixxi_fitcm_noB()}}.
}
\description{
Builds conditional-model deviance, score, and cross-derivative functions from
a prepared data list.
}
\seealso{
Other model fitting helpers:
\code{\link[=cM_devfun_xifix_datl_logT]{cM_devfun_xifix_datl_logT()}},
\code{\link[=fitWeibS]{fitWeibS()}},
\code{\link[=fixxi_fitcm0]{fixxi_fitcm0()}},
\code{\link[=fixxi_fitcm_noB]{fixxi_fitcm_noB()}},
\code{\link[=init_fixxi]{init_fixxi()}},
\code{\link[=init_fixxi_logT]{init_fixxi_logT()}},
\code{\link[=jm_fit_wmods]{jm_fit_wmods()}},
\code{\link[=lmm_fit_jm0]{lmm_fit_jm0()}},
\code{\link[=xifix_subX]{xifix_subX()}}
}
\concept{model fitting helpers}
