% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cm-fit.R
\name{cm_full_adjusted_scores}
\alias{cm_full_adjusted_scores}
\title{Compute full-data adjusted conditional-model scores}
\usage{
cm_full_adjusted_scores(fit1, fit_w, m_YS)
}
\arguments{
\item{fit1}{First-stage conditional-model fit with derivative components.}

\item{fit_w}{Fitted Weibull survival object.}

\item{m_YS}{Match index between longitudinal and survival score rows.}
}
\value{
Subject-level adjusted score matrix.
}
\description{
Compute full-data adjusted conditional-model scores
}
\keyword{internal}
