% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cm-fit.R
\name{cm_inverse_hessian_cov}
\alias{cm_inverse_hessian_cov}
\title{Compute an inverse-Hessian covariance matrix}
\usage{
cm_inverse_hessian_cov(hessian)
}
\arguments{
\item{hessian}{Numerical Hessian matrix.}
}
\value{
Inverse Hessian computed with Cholesky backsolves.
}
\description{
Compute an inverse-Hessian covariance matrix
}
\keyword{internal}
