% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cm-fit.R
\name{cm_sandwich_cov}
\alias{cm_sandwich_cov}
\title{Compute a sandwich covariance matrix from subject-level scores}
\usage{
cm_sandwich_cov(hessian, score_i)
}
\arguments{
\item{hessian}{Numerical Hessian matrix.}

\item{score_i}{Subject-level score matrix, one subject per row.}
}
\value{
Sandwich covariance matrix.
}
\description{
Compute a sandwich covariance matrix from subject-level scores
}
\keyword{internal}
