% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-wrappers.R
\name{cm_screen_biomarker_logT}
\alias{cm_screen_biomarker_logT}
\title{Fit one C++ log-time conditional-model screen for a high-dimensional biomarker}
\usage{
cm_screen_biomarker_logT(
  prepared,
  biomarker_id,
  whiten = TRUE,
  return_timing = FALSE
)
}
\arguments{
\item{prepared}{Prepared screening object returned by \code{\link[=prepare_biomarker_screen]{prepare_biomarker_screen()}}.}

\item{biomarker_id}{Column id in \code{sim_data$y} to screen.}

\item{whiten}{Whether to use the whitened optimizer. Defaults to \code{TRUE}.}

\item{return_timing}{Whether to return elapsed time in seconds for the
C++ CM starting-value step.}
}
\value{
Seven conditional-model statistics for one biomarker. If
\code{return_timing = TRUE}, returns a list with \code{values} and \code{init_time_sec}.
}
\description{
Log-time variant of \code{\link[=cm_screen_biomarker]{cm_screen_biomarker()}} where the third fixed-effect
coefficient multiplies \code{log(Ti)} instead of \code{Ti}.
}
\seealso{
Other screening workflows:
\code{\link[=cm_screen]{cm_screen()}},
\code{\link[=cm_screenR]{cm_screenR()}},
\code{\link[=cm_screen_biomarker]{cm_screen_biomarker()}},
\code{\link[=cm_screen_biomarkerR]{cm_screen_biomarkerR()}},
\code{\link[=cm_screen_biomarker_logTR]{cm_screen_biomarker_logTR()}},
\code{\link[=cm_screen_biomarkers]{cm_screen_biomarkers()}},
\code{\link[=cm_screen_biomarkersR]{cm_screen_biomarkersR()}},
\code{\link[=cm_screen_biomarkers_logT]{cm_screen_biomarkers_logT()}},
\code{\link[=cm_screen_biomarkers_logTR]{cm_screen_biomarkers_logTR()}},
\code{\link[=cm_screen_fitxi]{cm_screen_fitxi()}},
\code{\link[=cm_screen_fitxi_logT]{cm_screen_fitxi_logT()}},
\code{\link[=cm_screen_stats]{cm_screen_stats()}},
\code{\link[=cm_screen_stats_logT]{cm_screen_stats_logT()}},
\code{\link[=cox_screen_biomarkers]{cox_screen_biomarkers()}},
\code{\link[=cox_screen_pvalue]{cox_screen_pvalue()}},
\code{\link[=fit_screen_biomarkers]{fit_screen_biomarkers()}},
\code{\link[=fit_screen_biomarkers_logT]{fit_screen_biomarkers_logT()}},
\code{\link[=fit_screen_simulations]{fit_screen_simulations()}},
\code{\link[=fit_screen_simulations_logT]{fit_screen_simulations_logT()}},
\code{\link[=jm_screen_biomarkers]{jm_screen_biomarkers()}},
\code{\link[=jm_screen_pvalue]{jm_screen_pvalue()}},
\code{\link[=prepare_biomarker_screen]{prepare_biomarker_screen()}}
}
\concept{screening workflows}
