% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-wrappers.R
\name{cox_screen_pvalue}
\alias{cox_screen_pvalue}
\title{Fit the Cox screen to one simulated dataset}
\usage{
cox_screen_pvalue(dat)
}
\arguments{
\item{dat}{Simulated dataset list with \code{X} and \code{S} components.}
}
\value{
A numeric vector containing the Cox estimate, standard error, and p-value.
}
\description{
Advanced helper used by \code{\link[=fit_screen_simulations]{fit_screen_simulations()}} for low-dimensional
simulations.
}
\seealso{
Other screening workflows:
\code{\link[=cm_screen]{cm_screen()}},
\code{\link[=cm_screenR]{cm_screenR()}},
\code{\link[=cm_screen_biomarker]{cm_screen_biomarker()}},
\code{\link[=cm_screen_biomarkerR]{cm_screen_biomarkerR()}},
\code{\link[=cm_screen_biomarker_logT]{cm_screen_biomarker_logT()}},
\code{\link[=cm_screen_biomarker_logTR]{cm_screen_biomarker_logTR()}},
\code{\link[=cm_screen_biomarkers]{cm_screen_biomarkers()}},
\code{\link[=cm_screen_biomarkersR]{cm_screen_biomarkersR()}},
\code{\link[=cm_screen_biomarkers_logT]{cm_screen_biomarkers_logT()}},
\code{\link[=cm_screen_biomarkers_logTR]{cm_screen_biomarkers_logTR()}},
\code{\link[=cm_screen_fitxi]{cm_screen_fitxi()}},
\code{\link[=cm_screen_fitxi_logT]{cm_screen_fitxi_logT()}},
\code{\link[=cm_screen_stats]{cm_screen_stats()}},
\code{\link[=cm_screen_stats_logT]{cm_screen_stats_logT()}},
\code{\link[=cox_screen_biomarkers]{cox_screen_biomarkers()}},
\code{\link[=fit_screen_biomarkers]{fit_screen_biomarkers()}},
\code{\link[=fit_screen_biomarkers_logT]{fit_screen_biomarkers_logT()}},
\code{\link[=fit_screen_simulations]{fit_screen_simulations()}},
\code{\link[=fit_screen_simulations_logT]{fit_screen_simulations_logT()}},
\code{\link[=jm_screen_biomarkers]{jm_screen_biomarkers()}},
\code{\link[=jm_screen_pvalue]{jm_screen_pvalue()}},
\code{\link[=prepare_biomarker_screen]{prepare_biomarker_screen()}}
}
\concept{screening workflows}
