% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-primitives.R
\name{get_bIS}
\alias{get_bIS}
\title{Draw intercept, slope, and changepoint effects}
\usage{
get_bIS(N, sb, rho, sc = 0, ru = function(N) runif(N, 6, 10))
}
\arguments{
\item{N}{Number of subjects to simulate.}

\item{sb}{Random-effect standard deviations.}

\item{rho}{Random-effect correlation parameter.}

\item{sc}{Random changepoint slope standard deviation.}

\item{ru}{Function that draws changepoints.}
}
\value{
A list with \code{m} random intercept/slope effects and \code{mC}
changepoint effects.
}
\description{
Advanced simulation helper for subject-specific intercept, slope, and
changepoint effects.
}
\seealso{
Other simulation generators:
\code{\link[=CM_sim_RI]{CM_sim_RI()}},
\code{\link[=CM_sim_RI_logT]{CM_sim_RI_logT()}},
\code{\link[=Cox_sim_ISCP]{Cox_sim_ISCP()}},
\code{\link[=JM_sim_ISCP]{JM_sim_ISCP()}},
\code{\link[=LamCox_Xt]{LamCox_Xt()}},
\code{\link[=Lam_mt_IScp]{Lam_mt_IScp()}},
\code{\link[=Lam_tc]{Lam_tc()}},
\code{\link[=chol_IS2]{chol_IS2()}},
\code{\link[=gfun0]{gfun0()}},
\code{\link[=lHw]{lHw()}},
\code{\link[=lHwi]{lHwi()}},
\code{\link[=lh_w]{lh_w()}},
\code{\link[=mt_IScp_i]{mt_IScp_i()}},
\code{\link[=survY_jm]{survY_jm()}},
\code{\link[=surv_Cox]{surv_Cox()}}
}
\concept{simulation generators}
