% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-batch.R
\name{get_seed}
\alias{get_seed}
\title{Draw reproducible integer seeds}
\usage{
get_seed(n = 1, s0 = NULL)
}
\arguments{
\item{n}{Number of values to draw.}

\item{s0}{Optional seed used before drawing new seeds.}
}
\value{
Numeric vector of length \code{n}.
}
\description{
Draw integer seeds used to fan out reproducible simulation jobs.
}
\seealso{
Other batch helpers:
\code{\link[=batch_setup]{batch_setup()}},
\code{\link[=batch_task_grid]{batch_task_grid()}},
\code{\link[=padLeft]{padLeft()}},
\code{\link[=parse_args]{parse_args()}},
\code{\link[=replicate_seed_grid]{replicate_seed_grid()}},
\code{\link[=simulate_seed_reps]{simulate_seed_reps()}},
\code{\link[=simulate_seed_vec]{simulate_seed_vec()}}
}
\concept{batch helpers}
