% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cm-likelihood.R
\name{lc2_logT}
\alias{lc2_logT}
\title{Cached log-time censored-subject integration function}
\usage{
lc2_logT(th0, Xd0, Sfun, l_x = lxx_b0)
}
\description{
Cached log-time censored-subject integration function
}
\keyword{internal}
