% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cm-likelihood.R
\name{lc_fun_logT}
\alias{lc_fun_logT}
\title{Create a log-time censored-subject integration function}
\usage{
lc_fun_logT()
}
\value{
A function used for censored-subject likelihood integration in the
log-time conditional-model variant.
}
\description{
Create a log-time censored-subject integration function
}
\keyword{internal}
