% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-cm-likelihood.R
\name{lxx_b0}
\alias{lxx_b0}
\title{Evaluate the random-intercept longitudinal log likelihood}
\usage{
lxx_b0(r, sb, sig)
}
\arguments{
\item{r}{Richardson depth or residual vector, depending on context.}

\item{sb}{Random-effect standard deviations.}

\item{sig}{Input used by this helper.}
}
\value{
A value used by the longbioscreen simulation pipeline.
}
\description{
Helper retained for the longbioscreen simulation pipeline.
}
\keyword{internal}
