% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-primitives.R
\name{mt_IScp_i}
\alias{mt_IScp_i}
\title{Create a subject trajectory function}
\usage{
mt_IScp_i(B = c(0, 0), Btau = 0, b, bC)
}
\arguments{
\item{B}{Fixed longitudinal intercept and slope coefficients.}

\item{Btau}{Fixed changepoint or acute-effect coefficient.}

\item{b}{Subject-specific random intercept and slope values.}

\item{bC}{Subject-specific random changepoint values.}
}
\value{
A function of visit time \code{tim}.
}
\description{
Advanced simulation helper that returns a subject-specific biomarker
trajectory function.
}
\seealso{
Other simulation generators:
\code{\link[=CM_sim_RI]{CM_sim_RI()}},
\code{\link[=CM_sim_RI_logT]{CM_sim_RI_logT()}},
\code{\link[=Cox_sim_ISCP]{Cox_sim_ISCP()}},
\code{\link[=JM_sim_ISCP]{JM_sim_ISCP()}},
\code{\link[=LamCox_Xt]{LamCox_Xt()}},
\code{\link[=Lam_mt_IScp]{Lam_mt_IScp()}},
\code{\link[=Lam_tc]{Lam_tc()}},
\code{\link[=chol_IS2]{chol_IS2()}},
\code{\link[=get_bIS]{get_bIS()}},
\code{\link[=gfun0]{gfun0()}},
\code{\link[=lHw]{lHw()}},
\code{\link[=lHwi]{lHwi()}},
\code{\link[=lh_w]{lh_w()}},
\code{\link[=survY_jm]{survY_jm()}},
\code{\link[=surv_Cox]{surv_Cox()}}
}
\concept{simulation generators}
