% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-highdim.R
\name{rho_fun_blockK}
\alias{rho_fun_blockK}
\title{Create a block correlation matrix}
\usage{
rho_fun_blockK(p, rho, K = 5)
}
\arguments{
\item{p}{Matrix dimension.}

\item{rho}{Within-block correlation.}

\item{K}{Block size.}
}
\value{
A \code{p} by \code{p} correlation matrix.
}
\description{
Builds a correlation matrix with constant within-block correlation and zero
cross-block correlation.
}
\seealso{
Other high-dimensional simulation:
\code{\link[=CM_givT_IS]{CM_givT_IS()}},
\code{\link[=CM_givT_IS_logT]{CM_givT_IS_logT()}},
\code{\link[=default_drawY100_effects]{default_drawY100_effects()}},
\code{\link[=drawY100]{drawY100()}},
\code{\link[=drawY100_logT]{drawY100_logT()}},
\code{\link[=jmCox_draw_W]{jmCox_draw_W()}},
\code{\link[=jmCox_draw_WD]{jmCox_draw_WD()}},
\code{\link[=rho_fun_band]{rho_fun_band()}},
\code{\link[=surv_Cox_W]{surv_Cox_W()}}
}
\concept{high-dimensional simulation}
